+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: Add `SingleEntryTable` component.
+
 ## [2.68.0] - 2020-05-06
 
 Feature: